dom/workers/test/test_bug1132395.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for 1132395</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script class="testbody" type="text/javascript">

SpecialPowers.pushPrefEnv({ set: [["dom.workers.sharedWorkers.enabled", true]] }, function() {
  var sw = new SharedWorker('bug1132395_sharedWorker.js');
  sw.port.onmessage = function(event) {
     ok(true, "We didn't crash.");
    SimpleTest.finish();
  }
});

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
